backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
transformers==4.46.0
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
pillow==11.0.0
supabase==2.8.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
accelerate==1.1.1
huggingface-hub==0.26.1
